refactor(continuation): dedup resolveContinuationRuntimeConfig — continuation/config.ts survives
reply/continuation-runtime.ts was a near-exact copy missing clampDelayMs.
RFC §5 designates continuation/config.ts as the canonical config surface.
Delete the duplicate, retarget all callers (agent-runner, subagent-announce,
status-text, continue-delegate-tool, post-compaction-delegate-dispatch).
