+++ src/status/status-text.ts
@@ -17,9 +17,9 @@ import {
   pendingDelegateCount,
   stagedPostCompactionDelegateCount,
 } from "../auto-reply/continuation-delegate-store.js";
+import { resolveContinuationRuntimeConfig } from "../auto-reply/continuation/config.js";
 import { normalizeGroupActivation } from "../auto-reply/group-activation.js";
 import { resolveSelectedAndActiveModel } from "../auto-reply/model-runtime.js";
-import { resolveContinuationRuntimeConfig } from "../auto-reply/reply/continuation-runtime.js";
 import type { ThinkLevel } from "../auto-reply/thinking.js";
 import { toAgentModelListLike } from "../config/model-input.js";
 import type { SessionEntry } from "../config/sessions.js";
